backend/requirements/base.txt
# OneVault Platform - Core Dependencies
# ====================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# Security and Authentication
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Validation and Serialization
marshmallow==3.20.2
email-validator==2.1.0

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Utilities
click==8.1.7
rich==13.7.0
structlog==23.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Monitoring and Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing (included in base for development ease)
pytest==7.4.3
pytest-asyncio==0.22.0
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production WSGI/ASGI
gunicorn==21.2.0 